Fix up ESP8266 support.

+++ src/SwitchInput.cpp
@@ -195,7 +195,7 @@ void RotaryEncoder::changePrecision(uint16_t maxValue, int currentValue) {
 
 void RotaryEncoder::increment(int8_t incVal) {
 	if(incVal >= 0) {
-		currentReading = min(currentReading + incVal, maximumValue);
+		currentReading = min((uint16_t)(currentReading + incVal), maximumValue);
 		callback(currentReading);
 	}
 	else if(currentReading != 0 && currentReading < abs(incVal)) {
